Improvement of event handler performance in Stage
The event handler function was fetched upon every fired event. This is
now avoided by fetching the function once before adding the listener.
This is partly a refactoring but also changes the code, since changes of
the events in that[] are now considered only during the call of
_bindContentEvents. If an event is removed however, the previous code
would crash, since that['_' + event] would then be null. Performance
profiling showed slight advantage, from 20% spent to 18% spent inside
event handler. And the code is cleaner.