Update Stage.js
At this time, only single finger touch is supported. However, if the first pointer event the stage receives is a multi-touch event, the value of stage.pointerPos will be left undefined. This causes:
```
`TypeError: 'undefined' is not an object (evaluating 'pos.x')`
```
in all internal the touch event handers. 
I suggest always using the location of the first finger in this case.