Add in offset for packed sprites
Many sprite animation tools pack the sprites in such a way that the
size of the image is optimized.  In order to play the sprite properly
you need to change the x and y coordinate of the image so the sprite
plays properly.  Currently in Kineticjs there is no place to store this
offset data.  I added the offset data optionally to the sprite object
and changed the scene function to use it if it exists.