Removing unused value from return statement in _eachAncestorReverse()
A minor cleanup. When method _eachAncestorReverse() does not have
the top node argument, it returns "nothing" (undefined). Thus, the
true value returned when it has the top argument – and the current
Konva code never checks or uses that returned value – seems to be
just a side effect of code copy-paste or alike, done in the past.

Unless the function was planned to return a boolean value for a
a good reason (in which case we'd need to add a return value to
its end), let's remove the unused value from return.