Simplify the Boolean condition used in _useBufferCanvas()
The original Boolean condition duplicates most of the predicates
used in both parts of the "or" expression. It is enough to test
them just once, which also makes it easier to figure out what
the criterion for using the buffer canvas is. (An extra effect
is that the code becomes smaller and potentially quicker for the
JIT compiler to handle, but I don't expect much reduction here.)