Very minor optimization to _getCache()
This is just a minor change to avoid accessing the same property
in object this._cache twice inside _getCache() when its is already
set (defined). No big performance improvements probably, but given
that the cached values can be checked many times for each node
in the tree, I think it makes sense to do that small improvement.