Minor improvement to the implementation of Text._sceneFunc
As the context's translate() is additive, there is no need for
the two consecutive calls of it inside _sceneFunc(): The same
effect can be achieved with with just one call of translate().

A corresponding update was done in the unit test for Text that
had a pair of calls to translate() in the expected call dump.