When removing a child, set key "parent" to null, don't delete it
Structural changes to objects are relatively more expensive and
harder to optimize for modern JavaScript compilers, compared to
just changing object's key values. The parent property of nodes
is already set to null now, as part constructing the object.
Setting it back to null on removal / destroy of objects is more
consistent and a tiny little bit more efficient.