Set DD.node to null, items of Node#_cache to undefined, instead of deleting
In the footsteps of my previous commit. This one replaces the deletion
of DD.node with setting it to null.

In addition, the internal "implementation contract" of Node's cache in
Konva is that a value is not in the cache if the corresponding key lookup
in the cache returns undefined. Thus, the suggested approach of avoiding
the delete operator works well here too.