Support for fuzzed canvas browsers
Brave, Tor, Safari in private mode, and other browsers implement canvas fuzzing. To work around this, this PR strips the hit detection canvas rendering to 6 bits instead of 8 for the colors (but only if fuzzing is detected). In practice, this limits the hit canvas to just over 262k objects, but the current implementation relies on a random color generator, so one would likely run into an issue with color generation due to the random generator long before the 250k object number would be hit anyways. It would be better to switch to a predictable color generator instead to ensure smooth generation for large numbers of shapes, but as it hasn't been an issue leaving the random generator.

This implementation also switches the HitCanvas to an OffscreenCanvas instead of adding an element to the DOM. It's unclear from the documentation if OffscreenCanvas is also affected by fuzzing, but it might not be.

Fixes #1928
Fixes #1769
Fixes #1132
Fixes #1619
Fixes #1484
Fixes #1195
Fixes #1151