Fix bounding box calculation for bezier lines.
With a bezier line, the client rect was based directly on the positions
of the points as if it were a normal line. This would result in a bounding
box that doesn't correspond to the actual rendered curve, which in turn
causes a Transformer containing the line to look incorrect.

Fix this by calculating the extrema of the bezier curve on each axis and
use those to get the bounds of the actual rendered line.